pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "dials-rest"
description = "A REST API to a subset of DIALS features"
readme = "README.md"
version = "0.0.1"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "python-dateutil",
    "fastapi",
    "ispyb",
    'importlib-metadata; python_version<"3.8"',
    "python-jose",
    "uvicorn[standard]",
]

[project.scripts]
create-access-token = "dials_rest.cli.create_access_token:run"